move forkcondition and displayhardfork types to fork crate
